+++ components/src/core/Drawer/DrawerBody.tsx
@@ -1,6 +1,6 @@
 import React, { HTMLAttributes } from 'react';
 import PropTypes from 'prop-types';
-import makeStyles from '@material-ui/core/styles/makeStyles';
+import makeStyles from '@mui/styles/makeStyles';
 import { DrawerNavGroup, DrawerNavGroupProps } from './DrawerNavGroup';
 import { NavItemSharedStyleProps, NavItemSharedStylePropTypes, SharedStyleProps, SharedStylePropTypes } from './types';
 import { mergeStyleProp } from './utilities';
@@ -71,6 +71,7 @@ const DrawerBodyRender: React.ForwardRefRenderFunction<unknown, DrawerBodyProps>
                 const groupProps: DrawerNavGroupProps = child.props;
 
                 return (
+                    /* @ts-ignore issue in the @types/react@17.0.41: https://github.com/mui/material-ui/issues/31932 */
                     <DrawerNavGroup
                         key={`NavGroup_${index}`}
                         {...groupProps}