+++ components/src/core/Drawer/utilities.tsx
@@ -1,7 +1,7 @@
 import React, { ReactNode } from 'react';
 
 // Use the child prop if it exists, or inherit from the parent prop
-export const mergeStyleProp = <T extends unknown>(parentValue: T, childValue: T): T | undefined =>
+export const mergeStyleProp = <T,>(parentValue: T, childValue: T): T | undefined =>
     childValue !== undefined ? childValue : parentValue;
 
 export const findChildByType = (children: ReactNode, type: string[]): JSX.Element[] =>