+++ components/src/core/InfoListItem/InfoListItem.styles.tsx
@@ -1,4 +1,6 @@
-import { createStyles, makeStyles, Theme } from '@material-ui/core/styles';
+import { Theme } from '@mui/material/styles';
+import createStyles from '@mui/styles/createStyles';
+import makeStyles from '@mui/styles/makeStyles';
 import color from 'color';
 import * as Colors from '@brightlayer-ui/colors';
 import { InfoListItemProps } from './InfoListItem';
@@ -10,6 +12,7 @@ export type InfoListItemClasses = {
     icon?: string;
     info?: string;
     listItemText?: string;
+    listItemButtonRoot?: string;
     rightComponent?: string;
     separator?: string;
     statusStripe?: string;
@@ -68,6 +71,7 @@ export const useStyles = makeStyles<Theme, InfoListItemProps>((theme: Theme) =>
             '&:focus': {
                 outline: 'none',
             },
+            padding: (props) => (props.onClick && props.ripple ? 0 : undefined),
         },
         avatar: {
             backgroundColor: (props) => props.statusColor || Colors.black[500],
@@ -92,6 +96,10 @@ export const useStyles = makeStyles<Theme, InfoListItemProps>((theme: Theme) =>
             marginLeft: (props) => (props.leftComponent ? (theme.direction === 'rtl' ? 0 : theme.spacing(2)) : 0),
             marginRight: (props) => (props.leftComponent ? (theme.direction === 'rtl' ? theme.spacing(2) : 0) : 0),
         },
+        listItemButtonRoot: {
+            height: 'inherit',
+            width: 'inherit',
+        },
         icon: {
             color: (props) => getIconColor(props, theme),
             justifyContent: (props) => getIconAlignment(props),
@@ -105,7 +113,7 @@ export const useStyles = makeStyles<Theme, InfoListItemProps>((theme: Theme) =>
             fontWeight: 400,
             lineHeight: 1.3,
             color: (props) =>
-                props.fontColor || (theme.palette.type === 'dark' ? theme.palette.text.secondary : 'inherit'),
+                props.fontColor || (theme.palette.mode === 'dark' ? theme.palette.text.secondary : 'inherit'),
         },
         rightComponent: {
             flex: '0 0 auto',
@@ -120,7 +128,7 @@ export const useStyles = makeStyles<Theme, InfoListItemProps>((theme: Theme) =>
             display: 'inline-block',
             lineHeight: 1.3,
             color: 'inherit',
-            margin: `0 ${theme.spacing(0.5)}px`,
+            margin: `0 ${theme.spacing(0.5)}`,
         },
         statusStripe: {
             position: 'absolute',
@@ -135,7 +143,7 @@ export const useStyles = makeStyles<Theme, InfoListItemProps>((theme: Theme) =>
             fontWeight: 400,
             lineHeight: 1.3,
             color: (props) =>
-                props.fontColor || (theme.palette.type === 'dark' ? theme.palette.text.secondary : 'inherit'),
+                props.fontColor || (theme.palette.mode === 'dark' ? theme.palette.text.secondary : 'inherit'),
         },
         title: {
             fontWeight: 600,